+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.10.0] - Mar 03, 2022
+
+### Added
+
+* Add [tar_contains validator](/docs/validators/#tarcontainsvalidator) to validate contents of a tar file
+* Add [zip_contains validator](/docs/validators/#zipcontainsvalidator) to validate contents of a zip file
+* Add [tempfile plugin](/docs/plugins/#tempfileplugin) to create temporary files for a test
+* Add `init` field for containers. Containers can use it to specify a custom initialization time (default=2)
+
+
 ## [1.9.0] - Dec 26, 2021
 
 ### Added